test(gpu): extract CUDA runtime identity tests
